+++ CHANGELOG.md
@@ -7,8 +7,10 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
-<!-- ### Changed -->
 ### Added
+
+-   Added `--csv-file` flag which writes raw NxN results result table to a CSV
+    file. ([#88](https://github.com/Polymer/tachometer/issues/88)).
 
 -   During auto-sampling, the time remaining before the timeout will be hit is
     now displayed ([#107](https://github.com/Polymer/tachometer/issues/107)).